Finished first pass of OrderBook

+++ Application/Common/Interfaces/IStockClient.cs
@@ -5,5 +5,5 @@ namespace Application.Common.Interfaces;
 public interface IStockClient
 {
     Task<List<Stock>> GetStocksBySymbolAsync(string name);
-    Task<double> GetRealtimePrice(string symbol );
+    Task<decimal> GetRealtimePrice(string symbol);
 }
